Split tests into multiple files.
This is a first step to clean up the test suite. This commit splits the
tests into multiple files.

fixes #49
