+++ tests/tests_utils.c
@@ -1,0 +1,82 @@
+/* Copyright © 2016 Ingo Bürk
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * Except as contained in this notice, the names of the authors or their
+ * institutions shall not be used in advertising or otherwise to promote the
+ * sale, use or other dealings in this Software without prior written
+ * authorization from the authors.
+ */
+#include <stdlib.h>
+#include <stdio.h>
+#include <stdbool.h>
+#include <stdarg.h>
+#include <string.h>
+
+#include "tests_utils.h"
+
+bool check_strings(const char *expected, const char *actual, const char *format, ...) {
+    va_list ap;
+
+    if (expected == NULL && actual == NULL)
+        return false;
+
+    if (expected != NULL && actual != NULL && strcmp(expected, actual) == 0)
+        return false;
+
+    va_start(ap, format);
+    vfprintf(stderr, format, ap);
+    va_end(ap);
+    return true;
+}
+
+bool check_ints(const int expected, const int actual, const char *format, ...) {
+    va_list ap;
+
+    if (expected == actual)
+        return false;
+
+    va_start(ap, format);
+    vfprintf(stderr, format, ap);
+    va_end(ap);
+    return true;
+}
+
+bool check_longs(const long expected, const long actual, const char *format, ...) {
+    va_list ap;
+
+    if (expected == actual)
+        return false;
+
+    va_start(ap, format);
+    vfprintf(stderr, format, ap);
+    va_end(ap);
+    return true;
+}
+
+int check_database(xcb_xrm_database_t *database, const char *expected) {
+    bool err = false;
+    char *actual = xcb_xrm_database_to_string(database);
+
+    fprintf(stderr, "== Assert that database is correct.\n");
+    err |= check_strings(expected, actual, "Expected database <%s>, but found <%s>\n", expected, actual);
+
+    if (actual != NULL)
+        free(actual);
+    return err;
+}
